Remove `github.repository_owner` condition on ghcr.io publish
I originally put a condition that the workflow should only publish to
ghcr.io if the owner of the repository is "white-axe" so that if someone
forks this repository it wouldn't publish packages without their
permission.

Well I forgot that forked repositories have workflows disabled by
default. So there's really no reason for this at all.