Tag ghcr.io image as latest only if this is the latest release
When a new release is created on GitHub, the workflow would build the
image, upload it to ghcr.io and update the latest tag on ghcr.io to
point to the newly uploaded image.

It's possible for the chronologically newest release on GitHub to not be
the latest release. For example if I release version 2 of this image and
then a day later I decide to make a new patch release for version 1 of
this image, the version 2 release should count as the latest release
even though the new version 1 release is chronologically newest.

To fix this problem, the workflow now checks if the current release is
the latest release, and only updates the latest tag on ghcr.io if so.