New image builder
I've created a more advanced OCI image builder that allows for multiple
layers. It stores compressed layers in the Nix store, so if you change
something and then rebuild, only changed layers need to be compressed
again, saving large amounts of time that would otherwise be spent on
compression. This is in addition to the time saved when importing the
image into a container runtime during testing, by only requiring the
container runtime to import changed layers instead of having to reimport
the entire image every time even a minor change is made.