Use `pkgs.ffmpeg-headless` instead of `pkgs.ffmpeg`
The default version of FFmpeg from Nixpkgs indirectly depends on
libcamera at runtime:

    ffmpeg > SDL2 > pipewire > libcamera

That's a problem because libcamera is known to build irreproducibly:
libcamera's build system incorporates a random number generated at build
time into the libcamera binaries for a reason that appears to be
required by the functionality of libcamera.

We can't accept this, though, because that makes the container image
irreproducible, thereby threatening the reproducibility of the music and
art generated by the image.

That's why I've replaced the default version of FFmpeg with the headless
version, which doesn't depend on SDL2, removing the image's dependency
on libcamera.