Remove build-time dependency on KVM
The image builder artificially has a dependency on /dev/kvm for
absolutely no reason. I've tried building this image on a machine with
a working /dev/kvm and on a machine with a nonfunctional /dev/kvm and
the resulting image is the same. And yet the image builder throws an
error if /dev/kvm is not present.

I've patched out the KVM dependency using an overlay.