Remove fpu.patch for jack2
The two kinds of SIMD optimizations I'm trying to remove are those that
are enabled depending on feature detection on the CPU building the
application and those that are enabled depending on feature detection on
the CPU running the application. jack2's SIMD optimizations are enabled
based on compiler flags so they're totally fine. Besides, the current
compiler flags seem to only enable SSE (version 1, not later versions)
which is available on all x86-64 CPUs.